visual: extract config from vmenus.c to config.c
